refactor(cli): improve structure and readability
