+++ packages/cli/bin/protokit-cli.js
@@ -1,8 +1,8 @@
 #!/usr/bin/env node
-import { spawn } from 'child_process';
-import path from 'path';
-import { fileURLToPath } from 'url';
-import process from 'process';
+import { spawn } from "child_process";
+import path from "path";
+import { fileURLToPath } from "url";
+import process from "process";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -11,29 +11,32 @@ const args = process.argv.slice(2);
 
 if (!process.env.TS_NODE_LOADER_ACTIVE) {
   const env = { ...process.env };
-  env.TS_NODE_LOADER_ACTIVE = '1';
+  env.TS_NODE_LOADER_ACTIVE = "1";
   env.NODE_OPTIONS = [
-    env.NODE_OPTIONS || '',
-    '--loader ts-node/esm',
-    '--experimental-vm-modules',
-    '--experimental-wasm-modules',
-    '--es-module-specifier-resolution=node',
-    '--no-warnings'
-  ].join(' ');
+    env.NODE_OPTIONS || "",
+    "--loader ts-node/esm",
+    "--experimental-vm-modules",
+    "--experimental-wasm-modules",
+    "--es-module-specifier-resolution=node",
+    "--no-warnings",
+  ].join(" ");
 
-  const child = spawn('node', [path.join(__dirname, '../dist/index.js'), ...args], {
-    stdio: 'inherit',
-    env
-  });
+  const child = spawn(
+    "node",
+    [path.join(__dirname, "../dist/index.js"), ...args],
+    {
+      stdio: "inherit",
+      env,
+    }
+  );
 
-  child.on('exit', (code) => process.exit(code));
-  child.on('error', (err) => {
+  child.on("exit", (code) => process.exit(code));
+  child.on("error", (err) => {
     console.error(err);
     process.exit(1);
   });
-
 } else {
-  import(path.join(__dirname, '../dist/index.js')).catch((err) => {
+  import(path.join(__dirname, "../dist/index.js")).catch((err) => {
     console.error(err);
     process.exit(1);
   });
